+++ api/README.md
@@ -1,16 +1,14 @@
-<h1 align="center">
-    <img  alt="serverless start"  src="https://github.com/pimp-my-book/serverless-graphql-nodejs-starter/blob/master/serverless-starter.png">
- </h1>
 
-# Serverless GraphQL NodeJS Starter ![GitHub](https://img.shields.io/github/license/mashape/apistatus.svg)
 
-This is a basic boilerplate to kitstart your Serverless GraphQL projects with AWS Lambda. 
+# Serverless GraphQL TypeScript Starter ![GitHub](https://img.shields.io/github/license/mashape/apistatus.svg)
+
+This is a basic TypeScript boilerplate to kitstart your Serverless GraphQL projects with AWS Lambda. 
 
 It is powered by [`Apollo-Server-Lambda`](https://github.com/apollographql/apollo-server/tree/master/packages/apollo-server-lambda) and uses the [Serverless Framework](https://serverless.com/) to deploy it to AWS.
 
 This is hugely inspired by the starter that [Anomaly Innovations](https://github.com/AnomalyInnovations/serverless-nodejs-starter ) created. The project was cloned from it. 
 
-## Prerequistes 
+## Prerequistes 🚁
 
 * [An AWS Account](https://aws.amazon.com/)
 * [A configured AWS CLI](https://serverless.com/framework/docs/providers/aws/guide/credentials/)
@@ -19,12 +17,12 @@ This is hugely inspired by the starter that [Anomaly Innovations](https://github
 
 Note this project uses [Yarn](https://yarnpkg.com). 
 
-## How to get started
+## How to get started 🎮
 
 First clone the repo:
 
 ```
-$ serverless install --url https://github.com/pimp-my-book/serverless-graphql-nodejs-starter --name project-name
+$ serverless install --url https://github.com/AmoDinho/serverless-graphql-typescript-starter.git --name project-name
 ```
 
 Move into the project folder:
@@ -54,12 +52,12 @@ $ yarn deploy
 This should be a good start to be able to move on and do whatever you gotta do meet business requirements.  
 
 
-### How To Link Other AWS Services
+### How To Link Other AWS Services 🗽
 
 #### DynamoDB
 #### Cognito
 
 
-## Contributing
+## Contributing😀 
 
 Contributions are most welcome!!
