+++ api/src/resolvers/query.ts
@@ -1,0 +1,3 @@
+export const hello = (args: any, context: any) => {
+    return "Your GraphQL API is now LIVE!🎈 "
+}
