+++ api/tsconfig.json
@@ -1,0 +1,28 @@
+{
+    "compilerOptions": {
+        "target": "es5",
+        "module": "esnext",
+        "outDir": "./dist",
+        "moduleResolution": "node",
+        "esModuleInterop": true,
+        "noImplicitAny": false,
+        "allowSyntheticDefaultImports": true,
+        "lib": [
+            "es2016",
+            "esnext.asynciterable"
+        ],
+        "typeRoots": [
+            "node_modules/@types/graphql"
+        ],
+        "sourceMap": true,
+        "baseUrl": "."
+    },
+    "include": [
+        "./src/**/*"
+    ],
+    "exclude": [
+        "./node_modules",
+        ".serverless",
+        ".webpack"
+    ]
+}
