--- server/src/resolvers/index.js
@@ -1,7 +1,0 @@
-import {hello} from "./query";
-
-export const resolvers = {
-    Query: {
-        hello: (root, args, context) => hello(args, context)
-    }
-}
