--- server/src/resolvers/query.js
@@ -1,3 +1,0 @@
-export const hello = (args, context) => {
-    return "Your GraphQL API is now LIVE!🎈 "
-}
